requirements.txt
Flask
python-dotenv==1.0.0
altair==4
tiktoken==0.4.0
qdrant-client==1.5.4
langchain==0.0.316
openai==0.28.1